SFA: Utilize scale_exp parameter of GXSetIndTexMtx
