+++ src/StarFoxAdventures/render.ts
@@ -8,16 +8,16 @@ import { GfxRenderInst, GfxRenderInstList, GfxRenderInstManager } from "../gfx/r
 import { CameraController } from '../Camera';
 import { pushAntialiasingPostProcessPass, setBackbufferDescSimple, standardFullClearRenderPassDescriptor } from '../gfx/helpers/RenderGraphHelpers';
 import { GfxrAttachmentSlot, GfxrGraphBuilder, GfxrRenderTargetDescription, GfxrTemporalTexture } from '../gfx/render/GfxRenderGraph';
-import { Color, colorFromARGB8, colorFromRGBA, colorFromRGBA8, colorNewFromRGBA8, White } from '../Color';
+import { colorNewFromRGBA8, White } from '../Color';
 import { TextureMapping } from '../TextureHolder';
 import { nArray } from '../util';
 import { TDDraw } from '../SuperMarioGalaxy/DDraw';
 
 import { SFAAnimationController } from './animation';
-import { MaterialBase, MaterialFactory, getKonstColorSel, getTexGenSrc, getTexCoordID, getIndTexStageID, getIndTexMtxID, getKonstAlphaSel, MaterialRenderContext, getPostTexGenMatrix, makeOpaqueColorTextureDownscale2x, makeOpaqueDepthTextureDownscale2x } from './materials';
-import { interpS16, mat4FromRowMajor, mat4SetRowMajor, mat4SetValue, radsToAngle16, vecPitch } from './util';
+import { MaterialFactory, HeatShimmerMaterial } from './materials';
+import { radsToAngle16, vecPitch } from './util';
 import { DepthResampler } from './depthresampler';
-import { getMatrixAxisZ, lerp } from '../MathHelpers';
+import { getMatrixAxisZ } from '../MathHelpers';
 
 export interface SceneRenderContext {
     viewerInput: Viewer.ViewerRenderInput;
@@ -36,70 +36,6 @@ export interface SFARenderLists {
     heatShimmer: GfxRenderInstList;
     waters: GfxRenderInstList;
     furs: GfxRenderInstList;
-}
-
-class HeatShimmerMaterial extends MaterialBase {
-    protected rebuildInternal() {
-        const texMap0 = this.genTexMap(makeOpaqueColorTextureDownscale2x());
-        const texMap1 = this.genTexMap(makeOpaqueDepthTextureDownscale2x());
-        const texMap2 = this.genTexMap(this.factory.getWavyTexture());
-
-        const texCoord0 = this.genTexCoord(GX.TexGenType.MTX2x4, getTexGenSrc(texMap0));
-
-        const pttexmtx0 = this.genPostTexMtx((dst: mat4, matCtx: MaterialRenderContext) => {
-            mat4.fromScaling(dst, [7.0, 7.0, 1.0]);
-            mat4SetValue(dst, 0, 3, matCtx.sceneCtx.animController.envAnimValue0 * 10.0);
-            mat4SetValue(dst, 1, 3, -matCtx.sceneCtx.animController.envAnimValue1 * 10.0);
-        });
-        const texCoord1 = this.genTexCoord(GX.TexGenType.MTX3x4, getTexGenSrc(texMap0), undefined, undefined, getPostTexGenMatrix(pttexmtx0));
-
-        const k0 = this.genKonstColor((dst: Color) => {
-            colorFromRGBA(dst, 1.0, 1.0, 1.0, 0xfc/0xff);
-        });
-
-        const stage0 = this.genTevStage();
-        this.mb.setTevDirect(stage0.id);
-        this.setTevOrder(stage0, texCoord0, texMap1);
-        // Sample depth texture as if it were I8 (i.e. copy R to all channels)
-        const swap3: GX_Material.SwapTable = [GX.TevColorChan.R, GX.TevColorChan.R, GX.TevColorChan.R, GX.TevColorChan.R];
-        this.mb.setTevSwapMode(stage0.id, undefined, swap3);
-        this.mb.setTevKAlphaSel(stage0.id, getKonstAlphaSel(k0));
-        this.setTevColorFormula(stage0, GX.CC.ZERO, GX.CC.ZERO, GX.CC.ZERO, GX.CC.ZERO);
-        this.setTevAlphaFormula(stage0, GX.CA.KONST, GX.CA.ZERO, GX.CA.ZERO, GX.CA.TEXA, GX.TevOp.SUB, undefined, GX.TevScale.SCALE_4);
-
-        const indTexMtx0 = this.genIndTexMtx((dst: mat4, matCtx: MaterialRenderContext) => {
-            let s = 0.5 * Math.sin(3.142 * matCtx.sceneCtx.animController.envAnimValue0 * 10.0);
-            let c = 0.5 * Math.cos(3.142 * matCtx.sceneCtx.animController.envAnimValue0 * 10.0);
-            mat4SetRowMajor(dst,
-                c,   s,   0.0, 0.0,
-                -s,  c,   0.0, 0.0,
-                0.0, 0.0, 0.0, 0.0,
-                0.0, 0.0, 0.0, 1.0
-            );
-            mat4.multiplyScalar(dst, dst, 1 / 64); // scale_exp -6
-        });
-
-        const indStage0 = this.genIndTexStage();
-        this.setIndTexOrder(indStage0, texCoord1, texMap2);
-        this.mb.setIndTexScale(getIndTexStageID(indStage0), GX.IndTexScale._1, GX.IndTexScale._1);
-
-        const stage1 = this.genTevStage();
-        this.mb.setTevIndirect(stage1.id, getIndTexStageID(indStage0), GX.IndTexFormat._8, GX.IndTexBiasSel.STU, getIndTexMtxID(indTexMtx0), GX.IndTexWrap.OFF, GX.IndTexWrap.OFF, false, false, GX.IndTexAlphaSel.OFF);
-        this.setTevOrder(stage1, texCoord0, texMap0);
-        this.setTevColorFormula(stage1, GX.CC.TEXC, GX.CC.ZERO, GX.CC.ZERO, GX.CC.ZERO);
-        this.setTevAlphaFormula(stage1, GX.CA.ZERO, GX.CA.ZERO, GX.CA.ZERO, GX.CA.APREV, undefined, undefined, GX.TevScale.SCALE_4);
-
-        const stage2 = this.genTevStage();
-        this.mb.setTevDirect(stage2.id);
-        this.setTevOrder(stage2, undefined, undefined, GX.RasColorChannelID.COLOR0A0);
-        this.setTevColorFormula(stage2, GX.CC.ZERO, GX.CC.ZERO, GX.CC.ZERO, GX.CC.CPREV);
-        this.setTevAlphaFormula(stage2, GX.CA.ZERO, GX.CA.APREV, GX.CA.RASA, GX.CA.ZERO, undefined, undefined, GX.TevScale.SCALE_4);
-
-        this.mb.setChanCtrl(GX.ColorChannelID.COLOR0A0, false, GX.ColorSrc.REG, GX.ColorSrc.VTX, 0, GX.DiffuseFunction.NONE, GX.AttenuationFunction.NONE);
-        this.mb.setBlendMode(GX.BlendMode.BLEND, GX.BlendFactor.SRCALPHA, GX.BlendFactor.INVSRCALPHA);
-        this.mb.setZMode(true, GX.CompareType.LESS, false);
-        this.mb.setAlphaCompare(GX.CompareType.ALWAYS, 0, GX.AlphaOp.AND, GX.CompareType.ALWAYS, 0);
-    }
 }
 
 const scratchVec0 = vec3.create();
@@ -135,7 +71,7 @@ export class SFARenderer implements Viewer.SceneGfx {
 
     public materialFactory: MaterialFactory;
     private shimmerddraw = new TDDraw();
-    private enableHeatShimmer: boolean = true; // TODO: set by camera triggers
+    private enableHeatShimmer: boolean = false; // TODO: set by camera triggers
     private heatShimmerMaterial: HeatShimmerMaterial | undefined = undefined;
     // TODO: Merge GXMaterialHelperGfx into SFAMaterial
     private heatShimmerMaterialHelper: GXMaterialHelperGfx | undefined = undefined;
@@ -278,16 +214,16 @@ export class SFARenderer implements Viewer.SceneGfx {
         getMatrixAxisZ(cameraFwd, sceneCtx.viewerInput.camera.worldMatrix);
         vec3.negate(cameraFwd, cameraFwd);
         const camPitch16 = radsToAngle16(vecPitch(cameraFwd));
-        let factor;
+        let pitchFactor;
         if (camPitch16 < 0)
-            factor = ((((camPitch16 & 0xffff) >> 8) - 0xc0) * 4) & 0xfc;
+            pitchFactor = ((((camPitch16 & 0xffff) >> 8) - 0xc0) * 4) & 0xfc;
         else
-            factor = 0xff;
+            pitchFactor = 0xff;
 
         const strength = 0xff;
         // const strength = 0xff * ((Math.sin(sceneCtx.animController.envAnimValue0) + 1) / 2); // TODO: controlled by camera triggers
         const a1 = (strength * 0xff) >> 8;
-        const a0 = (factor * strength) >> 8;
+        const a0 = (pitchFactor * strength) >> 8;
 
         this.shimmerddraw.beginDraw();
         this.shimmerddraw.begin(GX.Command.DRAW_QUADS);
